fix: sec controller qol
